Refactor sign-in modal and component to improve authentication flow
